requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
numpy
joblib
scikit-learn
groq
